Move Annotation entity to Core
