+++ src/Wallabag/CoreBundle/DataFixtures/AnnotationFixtures.php
@@ -5,7 +5,7 @@ namespace Wallabag\CoreBundle\DataFixtures;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Common\DataFixtures\DependentFixtureInterface;
 use Doctrine\Persistence\ObjectManager;
-use Wallabag\AnnotationBundle\Entity\Annotation;
+use Wallabag\CoreBundle\Entity\Annotation;
 use Wallabag\CoreBundle\Entity\Entry;
 use Wallabag\UserBundle\DataFixtures\UserFixtures;
 use Wallabag\UserBundle\Entity\User;
